release 13
additions
- added Movement Fix (BedAura, WoolWars, KillAura, AntiFireball, Water Bucket)
- added AntiDebuff, replaces Potions
- Added Dolphin, Wool Wars, and made Trajectories more similar to b4
- added Silent Rotations to Aim Assist

changes, fixes, & others
- improved Jump Reset, more legit & actually jumps resets
- improved WTap, added more settings
- improved code readability and consistency
- improved scripting api, support for latest pug scripts
- fixed several bugs with KillAura
- added more settings to Water Bucket
- fixed a render bug with ItemESP
- added Respawn Check to BedWars module for bed whitelist
- fixed some FPS drops and render issues, especially with nametags
- changed blatant modules like BHop, Scaffold to JD's Raven Bs's version (outdated however)

+++ src/main/java/keystrokesmod/clickgui/components/impl/DescriptionComponent.java
@@ -9,11 +9,11 @@ import org.lwjgl.opengl.GL11;
 public class DescriptionComponent extends Component {
     public DescriptionSetting desc;
     private ModuleComponent p;
-    public int o;
-    public int x;
-    public int y;
+    public float o;
+    public float x;
+    public float y;
 
-    public DescriptionComponent(DescriptionSetting desc, ModuleComponent b, int o) {
+    public DescriptionComponent(DescriptionSetting desc, ModuleComponent b, float o) {
         this.desc = desc;
         this.p = b;
         this.x = b.categoryComponent.getX() + b.categoryComponent.getWidth();
@@ -28,7 +28,7 @@ public class DescriptionComponent extends Component {
         GL11.glPopMatrix();
     }
 
-    public void updateHeight(int n) {
+    public void updateHeight(float n) {
         this.o = n;
     }
 }
