+++ src/main/java/keystrokesmod/utility/profile/ProfileManager.java
@@ -24,10 +24,7 @@ import net.minecraftforge.common.MinecraftForge;
 import java.io.File;
 import java.io.FileReader;
 import java.io.FileWriter;
-import java.util.ArrayList;
-import java.util.Iterator;
-import java.util.List;
-import java.util.Map;
+import java.util.*;
 
 public class ProfileManager {
     public static Minecraft mc = Minecraft.getMinecraft();
@@ -238,8 +235,8 @@ public class ProfileManager {
                             String element = setting.getValue().getAsString();
                             String[] statesStr = element.split(",");
 
-                            int posX = Integer.parseInt(statesStr[0]);
-                            int posY = Integer.parseInt(statesStr[1]);
+                            float posX = Float.parseFloat(statesStr[0]);
+                            float posY = Float.parseFloat(statesStr[1]);
 
                             for (CategoryComponent c : ClickGui.categories) {
                                 if (c.category.name().equals(settingName)) {
